requirements.txt
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
gdown
visualpriors
pytorch_pretrained_vit
ftfy==6.1.1
regex
numpy==1.21.6
tabulate
matplotlib
argparse
transformers==4.19.2
git+https://github.com/rwightman/pytorch-image-models.git
moderngl>=5.6.4
objloader>=0.2.0
opencv-python>=3.4.3
Pillow>=5.3.0
PyOpenGL>=3.1.0
PyQt5>=5.11.3
pyrr>=0.9.2
scipy>=1.1.0